Subo Cambio en el Main activity

+++ app/src/main/java/com/example/gonzalo/proyectofinalandroid/Usuario.java
@@ -63,7 +63,7 @@ public class Usuario {
         return latitud;
     }
 
-    public void setLatitud(float latitud) {
+    public void setLatitud(double latitud) {
         this.latitud = latitud;
     }
 
@@ -71,7 +71,7 @@ public class Usuario {
         return longitud;
     }
 
-    public void setLongitud(float longitud) {
+    public void setLongitud(double longitud) {
         this.longitud = longitud;
     }
 }
